RAG1_WEB_HANDOFF/requirements.txt
# RAG1 Dependencies
# Install: pip install -r requirements_rag1.txt

# Core RAG
openai>=1.30
chromadb>=0.5
pymupdf>=1.24
pydantic>=2.0
python-dotenv>=1.0

# DICOM -> YOLO -> RAG1 pipeline
numpy>=1.23
pydicom>=3.0
Pillow>=10.0
torch>=2.0
torchvision>=0.15
ultralytics>=8.4
fastapi>=0.115
uvicorn>=0.30
